+++ server/out/s_challenges.js
@@ -1,35 +1,90 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.challenges = exports.Challenge = void 0;
+exports.challenges = exports.ChallengeData = exports.Timespan = exports.Challenge = void 0;
 const connection_1 = require("./connection");
 class Challenge {
-    constructor(name, desc, imgUrl, difficulty) {
+    constructor(id, name, desc, imgUrl, difficulty) {
+        this.id = id;
         this.name = name;
         this.desc = desc;
         this.imgUrl = imgUrl;
         this.difficulty = difficulty;
+        this.timespan = null;
+        this.hl = [];
     }
     name;
     desc;
     imgUrl;
     difficulty;
-    static from(data) {
-        let d = new Challenge(data.name, data.desc, data.imgUrl, data.difficulty);
+    id;
+    timespan;
+    hl;
+    static from(id, data) {
+        let d = new Challenge(id, data.name, data.desc, data.imgUrl, data.difficulty);
         let ok = Object.keys(data);
         for (const k of ok) {
             d[k] = data[k];
         }
         return d;
     }
+    get ongoing() {
+        let ongoing = false;
+        if (this.timespan) {
+            let now = new Date().getUTCDate();
+            if (new Date(this.timespan.end).getUTCDate() > now && now > new Date(this.timespan.start).getUTCDate())
+                ongoing = true;
+        }
+        return ongoing;
+    }
+    serializeGet() {
+        return {
+            id: this.id,
+            name: this.name,
+            desc: this.desc,
+            imgUrl: this.imgUrl,
+            difficulty: this.difficulty,
+            timespan: this.timespan,
+            ongoing: this.ongoing,
+            hl: this.hl
+        };
+    }
 }
 exports.Challenge = Challenge;
+class Timespan {
+    constructor(start, end) {
+        this.start = start;
+        this.end = end;
+    }
+    start;
+    end;
+}
+exports.Timespan = Timespan;
+class ChallengeData {
+    constructor() {
+        this.submissions = [];
+    }
+    submissions;
+    static async fromCID(cid) {
+        let str = await (0, connection_1.read)("../challenges/" + cid + ".json");
+        if (!str)
+            return;
+        let data = new ChallengeData();
+        return data;
+    }
+    serializeGet() {
+        return {
+            submissions: this.submissions
+        };
+    }
+}
+exports.ChallengeData = ChallengeData;
 exports.challenges = new Map();
 async function initChallenges() {
     let names = await (0, connection_1.readdir)("../challenges");
     if (!names)
         return;
-    for (const name of names) {
-        let dataStr = await (0, connection_1.read)("../challenges/" + name, "utf8");
+    for (const id of names) {
+        let dataStr = await (0, connection_1.read)("../challenges/" + id, "utf8");
         let data;
         if (!dataStr)
             continue;
@@ -39,7 +94,8 @@ async function initChallenges() {
         catch (e) {
             return;
         }
-        exports.challenges.set(name, Challenge.from(data));
+        let cid = id.substring(0, id.length - 5);
+        exports.challenges.set(cid, Challenge.from(cid, data));
     }
 }
 initChallenges();